Final Thoughts
Although the backend for the percentage didn’t particularly work out,
on the whole everyone learned a lot (and played a lot of pool+foosball)
and Lawrence created a rather impressive disease tracker.

Perhaps in the following weeks more work could be done to flesh out our
ideas, but what form this fleshing out will take is up to the team.

Hasta la vista.